py-work/pyproject.toml
[project]
name = "py-work"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "numpy>=1.24.4",
    "Pillow>=10.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
